+++ pytsite/reload/_console_command.py
@@ -1,4 +1,4 @@
-"""PytSite Reload Console Commands.
+"""PytSite Reload Console Command
 """
 from pytsite import console as _console
 from . import _api
@@ -9,20 +9,22 @@ __license__ = 'MIT'
 
 
 class Reload(_console.Command):
-    """Reload Command.
+    """Reload Console Command
     """
 
     @property
     def name(self) -> str:
-        """Get name of the command.
+        """Get name of the command
         """
         return 'reload'
 
     @property
     def description(self) -> str:
-        """Get description of the command.
+        """Get description of the command
         """
         return 'pytsite.reload@reload_console_command_description'
 
     def exec(self):
+        """Execute the command
+        """
         _api.reload()