+++ pytsite/reload/res/assets/js/pytsite-reload.js
@@ -1,7 +1,0 @@
-define(['pytsite-http-api'], function (httpApi) {
-    return {
-        reload: function () {
-            return httpApi.post('reload');
-        }
-    }
-});