+++ pytsite/CHANGELOG.md
@@ -1,12 +1,28 @@
 # PytSite 6 Changelog
 
 
+## 6.0.1 (2017-11-27)
+
+### Fixed
+
+- `mongodb`: invalid language resources IDs.
+- `plugman`: recursion detection.
+
+
+### Changed
+
+- `reload`: part of the code moved to the separate plugin.
+
+
+
 ## 6.0 (2017-11-24)
 
 ### Changed
+
 - `db` renamed to `mongodb`.
 
 ### Removed
+
 - `admin`, `assetman`, `auth`, `auth_http_api`, `auth_password`,
   `auth_profile`, `auth_settings`, `auth_storage_odm`, `auth_ui`,
   `auth_web`, `browser`, `ckeditor`, `feed`, `file`, `file_storage_odm`,